==> notebooks/02-parse-questions.ipynb <==
import re
from dataclasses import dataclass
from pydantic import BaseModel
from typing import List, Dict, Optional, Literal

class AnswerOption(BaseModel):
    letter: Literal['А', 'Б', 'В', 'Г']
    option_text: str

class Question(BaseModel):
    id: str
    question: str
    options: List[AnswerOption]
    answer: Literal['А', 'Б', 'В', 'Г']
    chapter : int


with open("/Users/alfa/Code/financial_assistant/data/raw/тест неквал инвестор вопросы.txt", 'r', encoding='utf-8-sig') as f:
    text = f.read()

id2chapter = {
    1: 'Покупка иностранных акций',
    2: 'Акции, не включенные в котировальные списки',
    3: 'Допуск к необеспеченным сделкам (маржинальная торговля)',
    4: 'Заключение договоров РЕПО',
    5: 'Опционы, фьючерсы, производные финансовые инструменты',
    6: 'Структурные облигации',
    7: 'Паи закрытых паевых инвестиционных фондов (ЗПИФ)',
    8: 'Облигации российских эмитентов, которым не присвоен рейтинг или он ниже уровня',
    9: 'Облигации иностранных эмитентов в валюте (еврооблигации) которым не присвоен рейтинг или он ниже нужного уровня',
    10: 'Облигации со структурным доходом',
    11: 'Вопросы для допуска к иностранным ETF'
}

chapter2id = {v:k for k,v in id2chapter.items()}

def src2questions(text, name, chapter):
    text = text.strip()
    q_num = name.split(':')[1].strip()
    lines = [l for l in text.strip().split('\n') if l.strip() != '']
    question_text = lines[0]
    correct_answer = lines[-1].split(': ')[-1].strip()

    options_text = lines[1:-1]

    options = []
    for opt_text in options_text:
        letter = opt_text[0]
        text = opt_text[3:].strip()
        options.append({
            'letter': letter.strip(),
            'option_text' : text.strip()
        })
        
    question = {
        'id' : q_num,
        'question' : question_text,
        'options' : options,
        'answer' : correct_answer,
        'chapter' : chapter
    }
    return question

chapters_src = re.split(r'Часть \d+\.\s[^\n]+', text)[1:]
chapter_names = re.findall(r'Часть \d+\.\s[^\n]+', text)

questions = []

for src, id_chapter in zip(chapters_src, id2chapter):
    q_src = re.split(r'Вопрос: \d+.\d+', src)[1:]
    q_names = re.findall(r'Вопрос: \d+.\d+', src)
    for qt, qn in zip(q_src, q_names):
        questions.append(src2questions(qt, qn, id_chapter))


questions

len(questions)

def check_validity(q):
    opt = [AnswerOption(**q['options'][i]) for i in range(4)]
    question = Question(options=opt, **{k:v for k,v in q.items() if k != 'options'})

validity = [check_validity(q) for q in questions]
