==> notebooks/04-rag.ipynb <==
import os
import pandas as pd
from tqdm.auto import tqdm
import json
import json_repair
from typing import Dict, List, Optional

import tiktoken

from langchain.docstore.document import Document
from langchain.vectorstores import FAISS
from langchain.embeddings.openai import OpenAIEmbeddings

from langchain.text_splitter import RecursiveCharacterTextSplitter

import warnings
warnings.filterwarnings('ignore')

VSE_GPT_API_KEY = ""
os.environ["OPENAI_API_KEY"] = VSE_GPT_API_KEY

# Data

adv_df = pd.read_csv("/Users/alfa/Code/financial_assistant/data/interim/alfa_invest_advanced_paragraphs.csv")
beg_df = pd.read_csv("/Users/alfa/Code/financial_assistant/data/interim/alfa_invest_begginer_paragraphs.csv")

adv_df.info(), beg_df.info()

adv_df.head()

beg_df.head()

## Analyze texts lengths

enc = tiktoken.get_encoding("cl100k_base")

get_tokens_num = lambda text: len(enc.encode(text))

adv_df['num_tokens'] = adv_df.text.apply(get_tokens_num)
beg_df['num_tokens'] = beg_df.text.apply(get_tokens_num)

adv_df.num_tokens.describe(), beg_df.num_tokens.describe()

## Chunkize

### Example

splitter = RecursiveCharacterTextSplitter(chunk_size=256, chunk_overlap=16, length_function=get_tokens_num)

splitter.split_text(adv_df.iloc[5]['text'])

[get_tokens_num(c) for c in splitter.split_text(adv_df.iloc[5]['text'])]

### Build all chunks

all_chunks = []

for idx, row in tqdm(adv_df.iterrows()):
    chunks = splitter.split_text(row.text)
    for chunk_id, chunk in enumerate(chunks):
        all_chunks.append(
            Document(
                page_content=chunk,
                metadata={
                    "chunk_id" : chunk_id,
                    "paragraph_id" : row.paragraph_id,
                    "article_id" : row.article_id
                }
            )
        )

len(all_chunks)

for idx, row in beg_df.iterrows():
    chunks = splitter.split_text(row.text)
    for chunk_id, chunk in enumerate(chunks):
        all_chunks.append(
            Document(
                page_content=chunk,
                metadata={
                    "chunk_id" : chunk_id,
                    "paragraph_id" : row.paragraph_id,
                    "article_id" : row.article_id
                }
            )
        )

len(all_chunks)

all_chunks[100]

# Vector database

embedding_model = OpenAIEmbeddings(
    model="text-embedding-3-small", 
    openai_api_base = "https://api.vsegpt.ru/v1/"
)

db = FAISS.from_documents(all_chunks, embedding_model)
db.save_local("/Users/alfa/Code/financial_assistant/data/processed/docs_db_index")


## Vector DB search

user_query = """
Ликвидность акции характеризует
А) Способность инвестора продать акцию с минимальными для него потерями в минимальный срок. 
Б) Разницу цены такой акции на разных торговых площадках. 
В) Вероятность погашения акции компанией – эмитентом. 
Г) Ни один из ответов не является правильным.
"""

db.similarity_search(user_query, 3)

# RAG

from openai import OpenAI
VSE_GPT_API_KEY = ""

client = OpenAI(
    api_key=VSE_GPT_API_KEY,
    base_url="https://api.vsegpt.ru/v1",
)

## Questions

with open('/Users/alfa/Code/financial_assistant/data/interim/test_qual_investor/questions.json', 'r') as f:
    questions = json.load(f)

questions[0]

with open('/Users/alfa/Code/financial_assistant/data/interim/test_qual_investor/chapters.json', 'r') as f:
    chapters = json.load(f)

chapters

## Prompt

with open("/Users/alfa/Code/financial_assistant/artifacts/prompts/system_v1.md", "r") as f:
    system_prompt = f.read()

print(system_prompt)

with open("/Users/alfa/Code/financial_assistant/artifacts/prompts/rag_v1.md", "r") as f:
    prompt_template = f.read()

print(prompt_template)

## Evaluation

def get_rag_response(
    question_dict : Dict, 
    prompt_template: str, 
    system_prompt: str, 
    db : FAISS,
    model="openai/gpt-4o-2024-11-20",
    sampling_params: Dict = {
        "max_tokens": 512,
        "temperature": 0.5,
        "top_p": 0.9
    }
):
    
    options = [
        opt["letter"] + '. ' + opt["option_text"]
        for opt in question_dict["options"]
    ]

    question_with_options_text = '\n'.join([question_dict['question']] + options)

    top_docs = db.similarity_search(question_with_options_text, 3)
    top_docs_texts = [f"{i}. {doc.page_content}" for i, doc in enumerate(top_docs)]
    context = '\n'.join(top_docs_texts)

    prompt = prompt_template.format(
        context=context,
        question=question_dict['question'], 
        option_1=options[0],
        option_2=options[1],
        option_3=options[2], 
        option_4=options[3],
    )

    # return prompt

    messages = [
        {"role": "system", "content": system_prompt},
        {"role": "user", "content": prompt},
        {"role": "assistant", "content": "{"}
    ]

    response = client.chat.completions.create(
        model=model,
        messages=messages,
        **sampling_params
    )

    return response

# GPT-4o-mini

results = []
responses = []
for question_dict in tqdm(questions):
    response = get_rag_response(
        question_dict,
        prompt_template,
        system_prompt,
        db,
        model="openai/gpt-4o-mini"
    )
    responses.append(response)
    try:
        response_text =  response.choices[0].message.content
        if '{' not in response_text[:5]:
            response_text = '{' + response_text
        response_dict = json_repair.loads(response_text)

        results.append(
            question_dict | {
                "llm_answer" : response_dict["answer"],
                "llm_reasoning" : response_dict["reasoning"]
            }
        )
    except Exception as e:
        results.append({})

res_df = pd.DataFrame(results).set_index("id")
res_df['correct'] = (res_df.answer == res_df.llm_answer).astype(int)
res_df.info()

res_df.head()

res_df.correct.sum(), res_df.correct.mean()

res_df.groupby("chapter")['correct'].mean()

res_df.to_csv('/Users/alfa/Code/financial_assistant/data/results/gpt4o-mini_rag.csv')

# GPT-4o

results = []
responses = []
for question_dict in tqdm(questions):
    response = get_rag_response(
        question_dict,
        prompt_template,
        system_prompt,
        db,
    )
    responses.append(response)
    try:
        response_text =  response.choices[0].message.content
        if '{' not in response_text[:5]:
            response_text = '{' + response_text
        response_dict = json_repair.loads(response_text)

        results.append(
            question_dict | {
                "llm_answer" : response_dict["answer"],
                "llm_reasoning" : response_dict["reasoning"]
            }
        )
    except Exception as e:
        results.append({})

res_df = pd.DataFrame(results).set_index("id")
res_df['correct'] = (res_df.answer == res_df.llm_answer).astype(int)
res_df.info()

res_df.head()

res_df.correct.sum(), res_df.correct.mean()

res_df.groupby("chapter")['correct'].mean()

res_df.to_csv('/Users/alfa/Code/financial_assistant/data/results/gpt4o_rag.csv')
