==> notebooks/05-res-analysis.ipynb <==
import pandas as pd
import json
import os
from tqdm.auto import tqdm

from langchain.vectorstores import FAISS
from langchain.embeddings.openai import OpenAIEmbeddings

VSE_GPT_API_KEY = ""
os.environ["OPENAI_API_KEY"] = VSE_GPT_API_KEY

with open('/Users/alfa/Code/financial_assistant/data/interim/test_qual_investor/questions.json', 'r') as f:
    questions = json.load(f)

with open('/Users/alfa/Code/financial_assistant/data/interim/test_qual_investor/chapters.json', 'r') as f:
    chapters = json.load(f)

norm_zs = pd.read_csv('/Users/alfa/Code/financial_assistant/data/results/gpt4o_zeroshot.csv', index_col=0)
mini_zs = pd.read_csv('/Users/alfa/Code/financial_assistant/data/results/gpt4o-mini_zeroshot.csv', index_col=0)

norm_rag = pd.read_csv('/Users/alfa/Code/financial_assistant/data/results/gpt4o_rag.csv', index_col=0)
mini_rag = pd.read_csv('/Users/alfa/Code/financial_assistant/data/results/gpt4o-mini_rag.csv', index_col=0)

norm_zs.options = norm_zs.options.apply(eval)
mini_zs.options = mini_zs.options.apply(eval)
norm_rag.options = norm_rag.options.apply(eval)
mini_rag.options = mini_rag.options.apply(eval)

norm_zs.info()

# Metrics

norm_zs.correct.mean()

norm_rag.correct.mean()

mini_zs.correct.mean()

mini_rag.correct.mean()

# Check different

norm_rag[(norm_zs.correct==0) & (norm_rag.correct==1)]

norm_zs[(norm_zs.correct==0) & (norm_rag.correct==1)]

mini_rag[(mini_zs.correct==0) & (mini_rag.correct==1)]

embedding_model = OpenAIEmbeddings(
    model="text-embedding-3-small", 
    openai_api_base = "https://api.vsegpt.ru/v1/"
)

db = FAISS.load_local("/Users/alfa/Code/financial_assistant/data/processed/docs_db_index", embedding_model)

def get_context(
    question_dict
):
    
    options = [
        opt["letter"] + '. ' + opt["option_text"]
        for opt in question_dict["options"]
    ]

    question_with_options_text = '\n'.join([question_dict['question']] + options)

    top_docs = db.similarity_search(question_with_options_text, 3)
    top_docs_texts = [f"{i}. {doc.page_content}" for i, doc in enumerate(top_docs)]
    context = '\n'.join(top_docs_texts)
    return question_with_options_text, context

norm_zs.loc[2.5].to_dict()

options, context = get_context(norm_zs.loc[2.5].to_dict())

print(f'Вопрос:\n{options}')
print(f'\nКонтекст:\n{context}\n')
print('\nВерный ответ:', norm_zs.loc[2.5].to_dict()['answer'])
print('Ответ без контекста:', norm_zs.loc[2.5].to_dict()['llm_answer'])
print('Ответ c контекстом:', norm_rag.loc[2.5].to_dict()['llm_answer'])
print('\nОбоснование без контекста:', norm_zs.loc[2.5].to_dict()['llm_reasoning'])
print('Обоснование c контекстом:', norm_rag.loc[2.5].to_dict()['llm_reasoning'])

options, context = get_context(norm_zs.loc[4.6].to_dict())

print(f'Вопрос:\n{options}')
print(f'\nКонтекст:\n{context}\n')
print('\nВерный ответ:', norm_zs.loc[4.6].to_dict()['answer'])
print('Ответ без контекста:', norm_zs.loc[4.6].to_dict()['llm_answer'])
print('Ответ c контекстом:', norm_rag.loc[4.6].to_dict()['llm_answer'])
print('\nОбоснование без контекста:', norm_zs.loc[4.6].to_dict()['llm_reasoning'])
print('Обоснование c контекстом:', norm_rag.loc[4.6].to_dict()['llm_reasoning'])
